% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-prediction-scales.r
\name{adjust_pred_scale}
\alias{adjust_pred_scale}
\title{Adjust scales of prediction}
\usage{
adjust_pred_scale(gbm_predictions, distribution_obj)
}
\arguments{
\item{gbm_predictions}{a matrix containing predictions generated from \code{\link{predict.GBMFit}}.}

\item{distribution_obj}{a GBMDist object used in the fitting a gbm model and generating the gbm_predictions.}
}
\value{
gbm_predictions scaled to the response of the distribution
}
\description{
S3 method that adjusts the scales of the predictor variables
}

