% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-dist-parameters.r
\name{check_dist_params}
\alias{check_dist_params}
\title{Check Distribution Parameters}
\usage{
check_dist_params(empty_obj, ...)
}
\arguments{
\item{empty_obj}{A skeleton distribution object (see \code{\link{empty_distribution()}}).}

\item{...}{Parameters defining how to create a full distribution object,
see \code{\link{gbm_dist}}.}
}
\value{
Warnings/Errors
}
\description{
Check if function parameters for creating distribution
object are of correct form.
}

