% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-response-values.r
\name{check_response_values}
\alias{check_response_values}
\title{Check Response Values}
\usage{
check_response_values(y, distribution_obj)
}
\arguments{
\item{distribution_obj}{a GBMDist object}

\item{y}{a matrix or vector of outcomes}
}
\description{
check responses are consistent with the distribution specified.
}

