% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-factors.r
\name{convert_factors}
\alias{convert_factors}
\title{Convert factors}
\usage{
convert_factor(gbm_data_obj)
}
\arguments{
\item{gbm_data_obj}{a gbm_data object}
}
\value{
gbm_data object with ordered/factor predictor variables appropriately converted
}
\description{
Function that converts factors or ordered predictors
to an appropriate numeric type.
}

