% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-cv-groups.r
\name{create_cv_groups}
\alias{create_cv_groups}
\alias{create_cv_groups.BernoulliGBMDist}
\alias{create_cv_groups.PairwiseGBMDist}
\alias{create_cv_groups.default}
\title{Create CV Groups}
\usage{
create_cv_groups(gbm_data_obj, gbm_dist_obj, train_params, cv_folds,
 cv_class_stratify, fold_id)

\method{create_cv_groups}{default}(gbm_data_obj, gbm_dist_obj, train_params,
  cv_folds, cv_class_stratify, fold_id)

\method{create_cv_groups}{BernoulliGBMDist}(gbm_data_obj, gbm_dist_obj,
  train_params, cv_folds, cv_class_stratify, fold_id)

\method{create_cv_groups}{PairwiseGBMDist}(gbm_data_obj, gbm_dist_obj,
  train_params, cv_folds, cv_class_stratify, fold_id)
}
\arguments{
\item{gbm_data_obj}{a GBMData object containing all of the data used to fit
a gbm model.}

\item{gbm_dist_obj}{a GBMDist object specifying the distribution and
any additional parameters needed.}

\item{train_params}{a GBMTrainParams object containing generic parameters
defining how the model should be trained.}

\item{cv_folds}{a positive integer specifying the number of folds to be used
in cross-validation of the gbm fit. If cv_folds > 1 then cross-validation is
performed; the default of cv_folds is 1.}

\item{cv_class_stratify}{a bool specifying whether or not to stratify via
response outcome. Currently only applies to "Bernoulli" distribution and
defaults to false.}

\item{fold_id}{An optional vector of values identifying what fold
each observation is in.}
}
\value{
a vector of numbers mapping each row of data to a cv fold.
}
\description{
Methods that create cv groups for gbm2 fitting.
Used internally, within \code{\link{gbm_cross_val}}, to split data and distribution
specific parameters into train and test folds.
}
\author{
James Hickey
}

