% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-distributions.r
\name{create_dist}
\alias{create_dist}
\title{Create distributions}
\usage{
create_dist(empty_obj, ...)
}
\arguments{
\item{empty_obj}{An empty distribution object of the correct class.}

\item{...}{extra parameters used to define the distribution object,
see \code{\link{gbm_dist}}.}
}
\value{
an appropriated gbm distribution object
}
\description{
Creation methods that take an empty distribution object
and builds on top of the skeleton.
}
\author{
James Hickey
}
