% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-strata.r
\name{create_strata}
\alias{create_strata}
\title{Create Strata}
\usage{
create_strata(gbm_data_obj, train_params distribution_obj)
}
\arguments{
\item{gbm_data_obj}{gbm data object}

\item{train_params}{gbm training parameters object}

\item{distribution_obj}{a gbm distribution object - the strata in this object will be
updated if they exist - this is only the case for the CoxPH model at this moment.}
}
\value{
updated distribution_obj
}
\description{
Function used to create to appropriate strata vectors -
currently only implemented for the CoxPH distribution.
}

