% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-groups.r
\name{determine_groups}
\alias{determine_groups}
\title{Set Up Groups}
\usage{
determine_groups(original_data, response, distribution_obj)
}
\arguments{
\item{original_data}{the original data used on the initial call to \code{\link{gbm2}}.}

\item{response}{the response data provided on the initial call to \code{\link{gbm2}}.}

\item{distribution_obj}{a GBMDist object created using gbm_dist.}
}
\value{
an updated distribution_object with the group, group_order and group_index updated
}
\description{
Function to set up groups - currently this only
applies to the pairwise distribution
}

