% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-accessors.r
\name{distribution_name}
\alias{distribution_name}
\title{What is the distribution name used here?}
\usage{
distribution_name(obj, ...)
}
\arguments{
\item{obj}{the object for which the distribution name is needed}

\item{...}{other parameters}
}
\value{
a string identifying the distribution
}
\description{
What is the distribution name used here?
}
