% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty-distribution.r
\name{empty_distribution}
\alias{empty_distribution}
\title{Empty Distribution
 
Creates an empty distribution object}
\usage{
empty_distribution(name)
}
\arguments{
\item{name}{A string specifying which empty distribution object to build.}
}
\value{
An empty distribution class - only the name defined.
}
\description{
Empty Distribution
 
Creates an empty distribution object
}

