% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-obs-in-fold.r
\name{extract_obs_in_fold}
\alias{extract_obs_in_fold}
\title{Extract observations in fold}
\usage{
extract_obs_in_fold(gbm_data_obj, gbm_dist_obj, train_params, cv_groups, fold_num, is_verbose)
}
\arguments{
\item{gbm_data_obj}{a GBMData object containing all of the data used to fit a gbm model.}

\item{gbm_dist_obj}{a GBMDist object specifying the distribution and any additional parameters needed.}

\item{train_params}{a GBMTrainParams object containing generic parameters defining how the model should be
trained.}

\item{cv_groups}{vector of integers specifying which folds each row of data belongs to.}

\item{fold_num}{integer (>=1) specifying which fold under consideration.}

\item{is_verbose}{if TRUE, will print out progress and performance of the fitting.}
}
\value{
list of input gbm objects with appropriately extracted observations
}
\description{
Extract the relevant observations to fit a CV fold
}

