% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm2.r
\name{gbm2}
\alias{gbm2}
\title{GBM2}
\usage{
gbm2(formula, distribution=gbm_dist("Gaussian", ...), data, weights, offset,
train_params=training_params(num_trees=100, interaction_depth=1, min_num_obs_in_node=10, 
shrinkage=0.001, bag_fraction=0.5, id=seq_len(nrow(data)), num_train=round(0.5 * nrow(data)), num_features) 
var_monotone=NULL, var_names=NULL, keep_gbm_data=FALSE, is_verbose=FALSE)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit.  The formula may include
an offset term (e.g. y~offset(n) + x).}

\item{distribution}{a GBMDist object specifying the distribution and any additional parameters needed.}

\item{data}{a data frame containing the variables in the model.  By default, the variables are taken from the 
environment.}

\item{weights}{optional vector of weights used in the fitting process.  These weights must be positive but 
need not be normalized.}

\item{offset}{optional vector specifying the model offset; must be positive.}

\item{train_params}{a GBMTrainParams object which specifies the parameters used in growing decision trees.}

\item{var_monotone}{optional vector, the same length as the number of predictors, indicating the relationship
each variable has with the outcome.  It have a monotone increasing (+1) or decreasing (-1) or an arbitrary relationship.}

\item{var_names}{a vector of strings of containing the names of the predictor variables.}

\item{keep_gbm_data}{a bool specifying whether or not the gbm_data object created in this method should be
stored in the results.}

\item{is_verbose}{if TRUE, gbm2 will print out progress and performance of the fit.}
}
\value{
a gbm2 object.
}
\description{
Fits generalized boosted regression models - new API.
}

