% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-cross-val.r
\name{gbm_cross_val}
\alias{gbm_cross_val}
\title{GBM Cross Validation}
\usage{
gbm_cross_validation(gbm_data_obj, gbm_dist_obj, train_params, 
var_container, cv_folds, cv_groups, is_verbose)
}
\arguments{
\item{gbm_data_obj}{a GBMData object containing all of the data used to fit a gbm model.}

\item{gbm_dist_obj}{a GBMDist object specifying the distribution and any additional parameters needed.}

\item{train_params}{a GBMTrainParams object containing generic parameters defining how the model should be
trained.}

\item{var_container}{a GBMVarCont object which defines the properties of the predictor variables in the data.}

\item{cv_folds}{a positive integer specifying the number of folds to be used in cross-validation of the gbm
fit. If cv_folds > 1 then cross-validation is performed; the default of cv_folds is 1.}

\item{cv_groups}{vector of integers specifying which row of data belongs to which cv_folds.}

\item{is_verbose}{if TRUE, will print out progress and performance of the fitting.}
}
\value{
a \code{GBMFit} object which contains appropriate CV info is requested.
}
\description{
Create fitted object - Cross validation is performed if \code{cv_folds > 1}, otherwise only the full model
is fitted.
}

