% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-cv-errors.r
\name{gbm_cv_errors.GBMCVFit}
\alias{gbm_cv_errors.GBMCVFit}
\title{GBM CV errors}
\usage{
gbm_cv_errors(gbm_cv_fit, cv_folds, cv_group)
}
\arguments{
\item{gbm_cv_fit}{a GBMCVFit object containing CV gbm models}

\item{cv_folds}{a positive integer specifying the number of folds to be used in cross-validation of the gbm
fit.}

\item{cv_group}{vector of integers specifying which row of data belongs to which cv_fold.}
}
\description{
Calculate cross validation errors
}

