% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-data.r
\name{gbm_data}
\alias{gbm_data}
\title{GBM data object
Object storing the data used in training a gbm model.}
\usage{
gbm_data(x, y, weights, offset)
}
\arguments{
\item{x}{matrix or data-frame of the predictor variables}

\item{y}{matrix or vector of response variables}

\item{weights}{vector containing the weights applied to each data row}

\item{offset}{vector containing the response variables offsets}
}
\value{
a gbm_data object
}
\description{
GBM data object
Object storing the data used in training a gbm model.
}
\author{
James Hickey
}

