% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-fit.r
\name{gbm_fit}
\alias{gbm_fit}
\title{GBM2 Fit}
\usage{
gbm_fit(gbm_data_obj, gbm_dist_obj, train_params, var_container, is_verbose)
}
\arguments{
\item{gbm_data_obj}{a GBMData object containing all of the data used to fit a gbm model.}

\item{gbm_dist_obj}{a GBMDist object specifying the distribution and any additional parameters needed.}

\item{train_params}{a GBMTrainParams object containing generic parameters defining how the model should be
trained.}

\item{var_container}{a GBMVarCont object which defines the properties of the predictor variables in the data.}

\item{is_verbose}{if TRUE, will print out progress and performance of the fitting.}
}
\value{
a fitted gbm object
}
\description{
Wrapper function for calling the C++ gbm function
}

