% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gbmCrossVal.R
\name{gbmCrossValErr}
\alias{gbmCrossValErr}
\title{Get the gbm cross-validation error}
\usage{
gbmCrossValErr(cv.models, cv.folds, cv.group, nTrain, n.trees)
}
\description{
Get the gbm cross-validation error
}

