% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-parallel.r
\name{gbmParallel}
\alias{gbmParallel}
\title{Control parallelization options}
\usage{
gbmParallel(n.threads = 1, arrayChunkSize = 1024)
}
\arguments{
\item{n.threads}{the number of threads to use (a positive
integer).  The number of cores on your computer is a
reasonable default.}

\item{arrayChunkSize}{the size of chunks to use in array scans.
Values that are too small result in a great deal of overhead;
The default of 1024 appears reasonable, but do experiment.}
}
\value{
an object of type \code{gbmParallel}
}
\description{
\code{gbm} uses openmp to parallelize its core algorithm, and the
details are controlled by this object.  As guidance, set
\code{n.threads} to the number of cores on your computer, and set
\code{arrayChunkSize} to a reasonable - not necessarily small -
size.
}

