% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbmt-data.R
\name{gbmt_fit_}
\alias{gbmt_fit_}
\title{Fit a gbm model to a gbm prep obj}
\usage{
gbmt_fit_(object)
}
\arguments{
\item{object}{GBMPrep object from \code{gbmt_data}}
}
\description{
Fit a gbm model to a gbm prep obj
}

