% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-misc.r
\name{get_misc}
\alias{get_misc}
\title{Get Misc Parameters}
\usage{
get_misc(distribution_obj)
}
\arguments{
\item{distribution_obj}{a GBMDist object}
}
\value{
a list of parameters
}
\description{
S3 method to get the "misc" objects from GBM distributions
}

