% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-accessors.r
\name{iteration_error}
\alias{iteration_error}
\title{Extract errors from GBMFit objects}
\usage{
iteration_error(obj, which = c("train", "valid", "cv"))
}
\arguments{
\item{obj}{a GBMFit object}

\item{which}{which error measure to extract}
}
\value{
a vector giving the error by tree
}
\description{
Extract errors from GBMFit objects
}
