% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf-plot.r
\name{perf_plot}
\alias{perf_plot}
\title{Performance Plots}
\usage{
perf_plot(gbm_fit_obj, best_iter, out_of_bag_curve, overlay, method, main)
}
\arguments{
\item{gbm_fit_obj}{a \code{GBMFit} created from an initial call to
\code{\link{gbm2}}.}

\item{best_iter}{iteration specifying the optimum number of iterations. This is determined in
\code{\link{gbm_perf}}.}

\item{out_of_bag_curve}{logical indicating whether to plot the out-of-bag performance
measures in a second plot.}

\item{overlay}{if TRUE and out_of_bag_curve=TRUE then a right y-axis is added to
the training and test error plot and the estimated cumulative improvement in
the loss function is plotted versus the iteration number.}

\item{method}{indicate the method used to estimate the optimal number of
boosting iterations. \code{method="OOB"} computes the out-of-bag estimate
and \code{method="test"} uses the test (or validation) dataset to compute an
out-of-sample estimate. \code{method="cv"} extracts the optimal number of
iterations using cross-validation if \code{gbm2} was called with \code{cv_folds}>1.}

\item{main}{the main title for the plot.}
}
\description{
Plot the performance of a GBMFit object produced by calling \code{\link{gbm2}}. This
function is used in \code{\link{gbm_perf}}.
}

