% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-perf.r
\name{plot.GBMTPerformance}
\alias{plot.GBMTPerformance}
\title{Plot GBM performance details}
\usage{
\method{plot}{GBMTPerformance}(x, out_of_bag_curve = FALSE, overlay = TRUE,
  main = "", ...)
}
\arguments{
\item{x}{a \code{GBMTPerformance} object (created by
\code{\link{gbmt_performance}})}

\item{out_of_bag_curve}{indicates whether to plot the out-of-bag
performance measures in a second plot.}

\item{overlay}{if TRUE and out_of_bag_curve=TRUE then a right
y-axis is added to the training and test error plot and the
estimated cumulative improvement in the loss function is plotted
versus the iteration number.}

\item{main}{the main title for the plot.}

\item{\dots}{currently ignored}
}
\description{
The train and validation error (in black and red respectively) are
plotted against the iteration number.  If the initial model was
built with cross-validation, the cross-validation error is shown
in green.
}
\details{
The scale of the error measurement, shown on the
left vertical axis, depends on the \code{distribution} argument
used in the initial call to \code{\link{gbmt}}.
}

