% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-cv-predict.r
\name{predict.GBMCVFit}
\alias{predict.GBMCVFit}
\title{Predictions for CV fitted GBM models}
\usage{
\method{predict}{GBMCVFit}(object, data, best_iter_cv, cv_group = NULL,
  ...)
}
\arguments{
\item{object}{a GBMCVFit object containing CV gbm models}

\item{data}{a GBMData object containing all of the data
used to fit a gbm model.}

\item{best_iter_cv}{number of trees with the smallest cv error}

\item{cv_group}{NULL or an integer vector indicating which model
should be used fro prediction.}

\item{\dots}{other options to \code{\link{predict.GBMFit}}}
}
\value{
either a vector or matrix of predictions.
}
\description{
Calculate predictions from cross validated generalized boosting
model from gbm2.
}
\details{
If cv_groups is NULL, the predictions from the models fitted to
each cv subset are returned as a matrix.  Otherwise this argument
should be an integer vector indicating which of the k-fold models
should be used to predict each observation.
}
\author{
James Hickey
}
