% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-predict.r
\name{predict.GBMFit}
\alias{predict.GBMFit}
\title{Predict method for GBM Model Fits}
\usage{
\method{predict}{GBMFit}(gbm_fit_obj, new_data, num_trees, type = "link",
  single_tree = FALSE, ...)
}
\arguments{
\item{gbm_fit_obj}{Object of class inheriting from (\code{GBMFit})}

\item{new_data}{Data frame of observations for which to make predictions}

\item{num_trees}{Number of trees used in the prediction. If \code{num_trees} is
a vector, predictions are returned for each iteration specified.}

\item{type}{The scale on which gbm makes the predictions}

\item{single_tree}{If \code{single_tree=TRUE} then \code{gbm_predict}
returns only the predictions from tree(s) \code{num_trees}}

\item{\dots}{further arguments passed to or from other methods}
}
\value{
Returns a vector of predictions. By default the predictions are on
the scale of f(x). For example, for the Bernoulli loss the returned value is
on the log odds scale, poisson loss on the log scale, and coxph is on the
log hazard scale.

If \code{type="response"} then \code{gbm2} converts back to the same scale as
the outcome. Currently the only effect this will have is returning
probabilities for bernoulli and expected counts for poisson. For the other
distributions "response" and "link" return the same.
}
\description{
Predicted values based on a generalized boosted model object - from gbm2
}
\seealso{
\code{\link{gbm2}}
}
\keyword{models}
\keyword{regression}

