% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-predict.r
\name{predict}
\alias{predict}
\alias{predict.GBMFit}
\title{Predict method for GBM Model Fits}
\usage{
predict(gbm_fit_obj, new_data, num_trees, type = "link",
  single_tree = FALSE, ...)

\method{predict}{GBMFit}(gbm_fit_obj, new_data, num_trees, type = "link",
  single_tree = FALSE, ...)
}
\arguments{
\item{gbm_fit_obj}{Object of class inheriting from \code{GBMFit}.}

\item{new_data}{Data frame of observations for which to make predictions}

\item{num_trees}{Number of trees used in the prediction. If \code{num_trees} is
a vector, predictions are returned for each iteration specified.}

\item{type}{The scale on which gbm makes the predictions}

\item{single_tree}{If \code{single_tree=TRUE} then \code{gbm_predict}
returns only the predictions from tree(s) \code{num_trees}}

\item{\dots}{further arguments passed to or from other methods}
}
\value{
Returns a vector of predictions. By default the predictions are on
the scale of f(x). For example, for the Bernoulli loss the returned value is
on the log odds scale, poisson loss on the log scale, and coxph is on the
log hazard scale.

If \code{type="response"} then \code{gbm2} converts back to the same scale as
the outcome. Currently the only effect this will have is returning
probabilities for bernoulli and expected counts for poisson. For the other
distributions "response" and "link" return the same.
}
\description{
Predicted values based on a generalized boosted model object - from gbm2
}
\details{
\code{gbm_predict} produces predicted values for each observation in a new_data_set
\code{new_data} using the first \code{num_trees} iterations of the boosting
sequence. If \code{num_trees} is a vector than the result is a matrix with
each column representing the predictions from gbm models with
\code{num_trees[1]} iterations, \code{num_trees[2]} iterations, and so on.

The predictions from \code{gbm2} do not include the offset term. The user may
add the value of the offset to the predicted value if desired.

If \code{gbm_fit_obj} was fit using \code{\link{gbm2}}, there will be no
\code{Terms} component. Therefore, the user has greater responsibility to
make sure that \code{new_data} is of the same format (order and number of
variables) as the one originally used to fit the model.
}
\seealso{
\code{\link{gbm2}}
}
\keyword{models}
\keyword{regression}

