% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor-order.r
\name{predictor_order}
\alias{predictor_order}
\title{Predictor Order}
\usage{
predictor_order(gbm_data_obj, train_params)
}
\arguments{
\item{gbm_data_obj}{a \code{GBMData} object.}

\item{train_params}{a \code{GBMTrainParams} object.}
}
\value{
an updated gbm_data_object that now contains the predictor variable order.
}
\description{
Function that calculates the predictor variable order for GBM
}

