% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calibrate.plot.R
\name{quantile.rug}
\alias{quantile.rug}
\title{Quantile rug plot}
\usage{
\method{quantile}{rug}(x, prob = (0:10)/10, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{prob}{the quantiles of x to mark on the x-axis.}

\item{...}{additional graphics parameters currently ignored.}
}
\value{
No return values
}
\description{
Marks the quantiles on the axes of the current plot.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
plot(x,y)
quantile.rug(x)
}
\author{
Greg Ridgeway \email{gregridgeway@gmail.com}
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link[stats]{quantile}},
\code{\link[base]{jitter}}, \code{\link[graphics]{rug}}.
}
\keyword{aplot}

