% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructGBMdata.R
\name{reconstructGBMdata}
\alias{reconstructGBMdata}
\title{Reconstruct a GBM's Source Data}
\usage{
reconstructGBMdata(x)
}
\arguments{
\item{x}{a \code{\link{gbm.object}} initially fit using \code{\link{gbm}}}
}
\value{
Returns a data used to fit the gbm in a format that can subsequently
be used for plots and summaries
}
\description{
Helper function to reconstitute the data for plots and summaries. This
function is not intended for the user to call directly.
}
\author{
Harry Southworth
}
\seealso{
\code{\link{gbm}}, \code{\link{gbm.object}}
}
\keyword{manip}

