% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative-influence.r
\name{relative_influence}
\alias{relative_influence}
\title{Methods for estimating relative influence}
\usage{
relative_influence(gbm_fit_obj, num_trees, rescale=FALSE, sort_it=FALSE)
}
\arguments{
\item{gbm_fit_obj}{a \code{GBMFit} object created from an initial call to
\code{\link{gbmt}}.}

\item{num_trees}{the number of trees to use for computations. If not provided,
the function will guess: if a test set was used in fitting, the number of
trees resulting in lowest test set error will be used; otherwise, if
cross-validation was performed, the number of trees resulting in lowest
cross-validation error will be used; otherwise, all trees will be used.}

\item{rescale}{whether or not the result should be scaled. Defaults to
\code{FALSE}.}

\item{sort_it}{whether or not the results should be (reverse) sorted.
Defaults to \code{FALSE}.}
}
\value{
By default, returns an unprocessed vector of estimated relative
influences. If the \code{rescale} and \code{sort} arguments are used,
returns a processed version of the same.
}
\description{
Helper functions for computing the relative influence of each variable in
the \code{GBMFit} object.
}
\details{
These functions offer the different
methods for computing the relative influence in \code{\link{summary.GBMFit}}.

\code{\link{relative_influence}} is the same as that
described in Friedman (2001).
\code{\link{permutation_relative_influence}} randomly permutes each
predictor variable at a time and computes the associated reduction in
predictive performance. This is similar to the variable importance measures
Breiman uses for random forests, but \code{gbmt} currently computes using the
entire training dataset (not the out-of-bag observations).
}
\author{
James Hickey, Greg Ridgeway \email{gregridgeway@gmail.com}
}
\references{
J.H. Friedman (2001). "Greedy Function Approximation: A Gradient
Boosting Machine," Annals of Statistics 29(5):1189-1232.

L. Breiman (2001).
\href{http://oz.berkeley.edu/users/breiman/randomforest2001.pdf}{Random
Forests}.
}
\seealso{
\code{\link{summary.GBMFit}}
}
\keyword{hplot}

