% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order-methods.r
\name{reorder_fit}
\alias{reorder_fit}
\title{Ordering methods}
\usage{
reorder_fit(gbm_fit_obj, distribution_obj)
}
\arguments{
\item{gbm_fit_obj}{a \code{GBMFit} object produced by a previous call to \code{gbm2} or \code{gbm2.fit}.}

\item{distributin_obj}{the \code{GBMDist} object used to produce the gbm_fit_obj.}
}
\value{
a \code{GBMFit} object with an appropriately order fit.
}
\description{
Series of functions that are used to re-order the fitted gbm model,
depending on what distribution was used in the initial call to \code{gbm2}.
Currently reordering only takes place for generalized boosted models produced
using the CoxPH or Pairwise distributions.
}

