% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-and-join.r
\name{split_and_join}
\alias{split_and_join}
\title{Rearrange training set and distribution data for CV}
\usage{
split_and_join(gbm_data_obj, train_params, rows_in_training, rows_in_fold)
}
\arguments{
\item{gbm_data_obj}{a GBMData  containing correctly ordered and validated data.}

\item{train_params}{a validated GBMTrainParams object}

\item{rows_in_training}{vector of logicals that determine what data rows are in the training set.}

\item{rows_in_fold}{vector of logicals indicating whether a row of training data is in the fold or not.}
}
\value{
gbm_data_obj with validation data moved to end of fields and recalculation of predictor ordering
}
\description{
Methods converts the gbm data and distribution parameters for CV fitting
}

