% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-old-gbm.r
\name{to_old_gbm}
\alias{to_old_gbm}
\title{Convert GBMFit to previous gbm object}
\usage{
to_old_gbm(gbm_fit_obj)
}
\arguments{
\item{gbm_fit_obj}{a \code{GBMFit} object produced by a call to \code{\link{gbmt}}.}
}
\value{
a \code{gbm} object of the form from Version 2.1 of the package.
}
\description{
Function that takes a \code{GBMFit} object produced from a call
to \code{\link{gbmt}} and converts it to the form of a gbm object from
Version 2.1 of the package.
}
\author{
James Hickey
}

