% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-accessors.R
\name{trees}
\alias{trees}
\title{Extract trees from GBMFit objects}
\usage{
trees(obj)
}
\arguments{
\item{obj}{a GBMFit object}
}
\value{
a vector containing the trees generated in the fit
}
\description{
Extract trees from GBMFit objects
}

