% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-fold-dist-data.r
\name{update_fold_dist_data}
\alias{update_fold_dist_data}
\title{Recreate distribution data for CV folds}
\usage{
update_fold_dist_data(gbm_dist_obj, gbm_data_obj, train_params, rows_in_training, rows_in_fold)
}
\arguments{
\item{gbm_dist_obj}{a GBMDist object properly formatted for gbm_fit.}

\item{gbm_data_obj}{a GBMData object containing correctly ordered and validated data.}

\item{train_params}{a validated GBMTrainParams object}

\item{rows_in_training}{vector of logicals that determine what data rows are in the training set.}

\item{rows_in_fold}{vector of logicals indicating whether a row of training data is in the fold or not.}
}
\value{
modified \code{GBMDist} object for CV fit.
}
\description{
Modify distribution object data for CV gbm fit.
}

