% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gbm-data.r
\name{validate_gbm_data}
\alias{validate_gbm_data}
\title{Validate data}
\usage{
validata_gbm_data(gbm_data_obj, distribution_obj)
}
\arguments{
\item{gbm_data_obj}{A GBMData object created using the function "gbm_data".}

\item{distribution_obj}{A GBMDist object storing all the parameters which define the model
distribution selected.}
}
\value{
a validated gbm_data_obj
}
\description{
Validates that the data is consistent with the distribution selected.
}

