% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-properties.r
\name{var_container}
\alias{var_container}
\title{Variable Container}
\usage{
var_container(gbm_data_obj,  var_monotone=NULL, var_names=NULL)
}
\arguments{
\item{gbm_data_obj}{an initialized GBMData object for the fit}

\item{var_monotone}{an optional vector, the same length as the
number of predictors, indicating which variables have a monotone
increasing (+1), decreasing (-1), or arbitrary (0) relationship
with the outcome.}

\item{var_names}{a vector of strings of the same length as the 
1st dimension of the response.}
}
\value{
a \code{GBMVarCont} object
}
\description{
Contains properties associated with predictor variables.
}

