% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify-data.r
\name{verify_data}
\alias{verify_data}
\title{Verify data}
\usage{
verify-data(x, y)
}
\arguments{
\item{x}{the matrix or data-frame containing the predictive variables}

\item{y}{the response vector}
}
\value{
NA - Throws an error if data is of incorrect type
}
\description{
Functions used to determine whether the data input is valid - checks that
the predictor and response vectors are of the right form
}

