% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available-distributions.r
\name{available_distributions}
\alias{available_distributions}
\title{Available Distributions}
\usage{
available_distributions()
}
\value{
a vector containing the names of the available distributions in the package.
}
\description{
Returns a vector of the distributions implemented in the package.
}
\author{
James Hickey
}

