% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine-groups.r
\name{determine_groups}
\alias{determine_groups}
\title{Set Up Groups}
\usage{
determine_groups(col_names, gbm_data, distribution_obj)
}
\arguments{
\item{col_names}{vector of variable names used in fittig - required for Pairwise}

\item{gbm_data}{a GBMData object created using}

\item{distribution_obj}{a GBMDist object created using gbm_dist}
}
\value{
an updated distribution_object
}
\description{
Function to set up groups - currently this only
applies to the pairwise distribution
}

