% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-package.r
\name{gbm-internal}
\alias{gbm-internal}
\title{gbm internal functions}
\description{
Helper functions for preprocessing data prior to building the model
}
\details{
These are functions used internally by \code{gbmt} and not intended for
direct use by the user.
}

