% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-package.R
\name{gbm-internal}
\alias{checkID}
\alias{checkMissing}
\alias{checkOffset}
\alias{checkWeights}
\alias{gbm-internal}
\alias{gbmCluster}
\alias{getCVgroup}
\alias{getStratify}
\alias{getVarNames}
\alias{guessDist}
\title{gbm internal functions}
\arguments{
\item{y}{The response variable}

\item{d, distribution}{The distribution, either specified by the user or
implied}

\item{class.stratify.cv}{Whether or not to stratify, if provided by the user}

\item{i.train}{Computed internally by \code{gbm}}

\item{group}{The group, if using \code{distribution='pairwise'}}

\item{strat}{Whether or not to stratify}

\item{cv.folds}{The number of cross-validation folds}

\item{x}{The design matrix}

\item{id}{The interaction depth}

\item{w}{The weights}

\item{n}{The number of cores to use in the cluster.}

\item{o}{The offset}
}
\description{
Helper functions for preprocessing data prior to building the model
}
\details{
These are functions used internally by \code{gbm} and not intended for
direct use by the user.
}

