% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-parallel.r
\name{gbmParallel}
\alias{gbmParallel}
\title{Control parallelization options}
\usage{
gbmParallel(num_threads = 1, array_chunk_size = 1024)
}
\arguments{
\item{num_threads}{the number of threads to use (a positive
integer).  The number of cores on your computer is a
reasonable default.}

\item{array_chunk_size}{the size of chunks to use in array scans.
Values that are too small result in a great deal of overhead;
The default of 1024 appears reasonable, but do experiment.}
}
\value{
an object of type \code{gbmParallel}
}
\description{
\code{gbmt} uses openmp to parallelize its core algorithm, and the
details are controlled by this object.  As guidance, set
\code{num_threads} to the number of cores on your computer, and set
\code{array_chunk_size} to a reasonable - not necessarily small -
size.
}

