% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-misc.r
\name{get_misc}
\alias{get_misc}
\title{Get Misc Parameters}
\usage{
get_misc(distribution_obj)
}
\arguments{
\item{distribution_obj}{a GBMDist object}
}
\value{
a list of additional parameters specific to the distribution
}
\description{
S3 method to get the "misc" objects from GBM distributions.
This method (internal) is used with gbm_call to extract the appropriate "misc"
for passing to the C++ layer.
}

