% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-cv-predict.r
\name{predict.GBMCVFit}
\alias{predict.GBMCVFit}
\title{Predictions for CV fitted GBM models}
\usage{
\method{predict}{GBMCVFit}(object, data, best_iter_cv, cv_all = FALSE,
  ...)
}
\arguments{
\item{object}{a GBMCVFit object containing CV gbm models}

\item{best_iter_cv}{number of trees with the smallest cv error}

\item{cv_all}{return a matrix of predictions for all observations
or just a vector of predictions for the excluded observations.}

\item{\dots}{not currently used}

\item{gbm_data_obj}{a GBMData object containing all of the data
used to fit a gbm model.}
}
\value{
a matrix of predictions for each cv fold.
}
\description{
Calculate predictions from cross validated generalized boosting
model from gbm2.
}
\author{
James Hickey
}
